Save, change position of error handeling to Error Service.
