+++ BusinessLogic/Services/TestService.cs
@@ -1,10 +1,10 @@
 ﻿using BusinessLogic.IRepositories;
 using BusinessLogic.Projections;
-using UserInterface.Stores;
 using BusinessLogic.IModels;
 using BusinessLogic.Enums;
 using BusinessLogic.Models;
 using BusinessLogic.Classes;
+using UserInterface.Stores;
 
 namespace BusinessLogic.Services;
 
@@ -16,6 +16,7 @@ public class TestService
 
     public TestService(ITestRepository testRepository) => this.testRepository = testRepository;
 
+    #region Test Retrieval
     public List<Test> GetAllTests() => testRepository.GetAllTests();
 
     public Test GetTestById(Guid id) => testRepository.GetTestById(id);
@@ -23,23 +24,34 @@ public class TestService
     public List<TestProjection>? GetTestProjectionsByTargetAudienceId(Guid id) => testRepository.GetTestProjectionsByTargetAudienceId(id);
 
     public Test? GetTestByTargetAudienceIdAndActive(Guid targetAudienceId) => testRepository.GetTestByTargetAudienceIdAndActive(targetAudienceId);
+    #endregion
+
+    #region CRUD Test
+    public Test CreateTest()
+    {
+        test = testRepository.CreateTest();
+        return test;
+    }
 
     public void SaveTest(Test test) => testRepository.SaveTest(test);
 
     public void DeleteTest(Test test) => testRepository.DeleteTest(test);
 
     public void UpdateTest(Test test) => testRepository.UpdateTest(test);
+    #endregion
 
-    public Test CreateTest()
-    {
-        test = testRepository.CreateTest();
 
-        return test;
-    }
-    public void SetTest(Test test)
+    public void SetTest(Test test) => this.test = test;
+
+    public void SaveOrUpdateTest(Test test, bool newTest)
     {
-        this.test = test;
+        if (newTest)
+            SaveTest(test);
+        else
+            UpdateTest(test);
     }
+
+    #region ???? not sure yet TODO
     public TextQuestionOption ConvertStringToQuestionOption(string text, Guid textQuestionId)
     {
         return new TextQuestionOption { Id = new Guid(), Option = text, TextQuestionId = textQuestionId };
@@ -59,20 +71,16 @@ public class TestService
     {
         return questionOptions.Select(ConvertQuestionOptionToString).ToList();
     }
+    #endregion
 
-    public void SaveOrUpdateTest(Test test, bool newTest)
-    {
-        if (newTest)
-            SaveTest(test);
-        else
-            UpdateTest(test);
-    }
+    #region Questions
     public ToneAudiometryQuestion CreateToneAudiometryQuestion()
     {
         ToneAudiometryQuestion audioQuestion = new ToneAudiometryQuestion { Id = new Guid(), QuestionType = QuestionType.AudioQuestion };
         audioQuestion.QuestionNumber = GetNewHighestQuestionNumber(test, QuestionType.AudioQuestion);
         return audioQuestion;
     }
+
     public TextQuestion CreateTextQuestion()
     {
         TextQuestion textQuestion = new TextQuestion
@@ -90,12 +98,15 @@ public class TestService
         test.TextQuestions.Add(textQuestion);
         return test.TextQuestions.ToList();
     }
+
     public List<ToneAudiometryQuestion> AddToneAudiometryQuestion(ToneAudiometryQuestion audioQuestion)
     {
         test.ToneAudiometryQuestions.Add(audioQuestion);
         return test.ToneAudiometryQuestions.ToList();
     }
+    #endregion
 
+    #region Test Manipulation 
     public void ToggleActiveStatus(Guid id)
     {
         Test? test = GetTestById(id);
@@ -138,14 +149,14 @@ public class TestService
 
     public static int GetQuestionNumberIndex<T>(List<T> questions, int questionNumber) where T : IQuestion
     {
-        AssertQuestions(questions);
+        ErrorService.AssertQuestions(questions);
 
         return questions.FindIndex(q => q.QuestionNumber == questionNumber);
     }
 
     public static List<T> ShiftQuestionNumbers<T>(List<T> questions) where T : IQuestion
     {
-        AssertQuestions(questions);
+        ErrorService.AssertQuestions(questions);
         int newNumber = 1;
 
         foreach (IQuestion question in questions)
@@ -158,13 +169,12 @@ public class TestService
 
     public List<T> UpdateQuestion<T>(List<T> questions, int questionNumber, T question) where T : IQuestion
     {
-        AssertQuestions(questions);
+        ErrorService.AssertQuestions(questions);
         int index = GetQuestionNumberIndex(questions, questionNumber);
         if (index != -1)
         {
             questions[index] = question;
         }
-
         switch (question)
         {
             case TextQuestion:
@@ -174,160 +184,12 @@ public class TestService
                 test.ToneAudiometryQuestions = questions.Cast<ToneAudiometryQuestion>().ToList(); ;
                 break;
         }
-
         return questions;
-    }
-    public static int ParseStringToInt(string str)
-    {
-        int parsedInt;
-        bool success = int.TryParse(str, out parsedInt);
-
-        if (success)
-            return parsedInt;
-        else
-            return -1;
-
-    }
-    public static string ValidateDecibels(string StartingDecibelsString)
-    {
-        try
-        {
-            int startingDecibels = ParseStringToInt(StartingDecibelsString);
-
-            if (startingDecibels == -1)
-            {
-                return ErrorStore.ErrorNotValidInteger;
-            }
-
-            if (!TestService.IsValidDecibel(startingDecibels))
-            {
-                return ErrorStore.ErrorStartingDecibels;
-            }
-
-            return string.Empty; // No error, return empty string
-        }
-        catch (Exception ex)
-        {
-            // Log or handle the exception as needed
-            return "Er is wat misgegaan"; // Error message for exception
-        }
-    }
-
-    public static string ValidateFrequency(string FrequencyString)
-    {
-        try
-        {
-            int frequency = ParseStringToInt(FrequencyString);
-
-            if (frequency == -1)
-            {
-                return ErrorStore.ErrorNotValidInteger;
-            }
-
-            if (!TestService.IsValidHz(frequency))
-            {
-                return ErrorStore.ErrorFrequencyLimit;
-            }
-
-            return string.Empty; // No error, return empty string
-        }
-        catch (Exception ex)
-        {
-            // Log or handle the exception as needed
-            return "Er is wat misgegaan"; // Error message for exception
-        }
-    }
-      
-    public static string ValidateInput(string columnName, params object[] values)
-    {
-        string validationMessage = string.Empty;
-
-        switch (columnName)
-        {
-            case "TestName":
-                validationMessage = ValidateTestName((string)values[0]);
-                break;
-            case "Audience":
-                validationMessage = ValidateAudience((TargetAudience)values[0]);
-                break;
-            case "TestQuestion":
-                validationMessage = ValidateTestQuestion((string)values[0]);
-                break;
-            case "MultipleChoice":
-            case "HasInputField":
-                validationMessage = ValidateQuestionType((bool)values[0], (bool)values[1], (List<string>)values[2]);
-                break;
-            case "Frequency":
-                validationMessage = ValidateFrequency((string)values[0]);
-                break;
-            case "StartingDecibelsString":
-                validationMessage = ValidateDecibels((string)values[0]);
-                break;
-            default:
-                break;
-        }
-
-        return validationMessage;
-    }
-
-
-    public static string ValidateTestName(string str)
-    {
-        // Validate if the test name is empty or contains illegal characters
-        if (TestService.IsEmptyString(str!))
-            return ErrorStore.ErrorTestName;
-        else if (str.Contains(ErrorStore.IllegalCharacters))
-            return ErrorStore.ErrorIllegalCharacters;
-        return string.Empty;
-    }
-    public static string ValidateAudience(TargetAudience targetAudience)
-    {
-        // Validate if the audience is correctly selected
-        if (targetAudience == null || TestService.IsEmptyString(targetAudience.Label))
-            return ErrorStore.ErrorAudience;
-        return string.Empty;
-    }
-
-    public static string ValidateTestQuestion(string question)
-    {
-        if (IsEmptyString(question))
-            return ErrorStore.ErrorTestQuestion;
-        else if (ContatinsInvalidCharacters(question))
-            return ErrorStore.ErrorIllegalCharacters;
-        return string.Empty;
-    }
-    public static string ValidateQuestionType(bool inputField, bool multipleChoice, List<string> options)
-    {
-        if (!inputField && !multipleChoice)
-            return ErrorStore.ErrorQuestionAnwserType;
-        if (multipleChoice && options.Count < 2)
-            return ErrorStore.ErrorMultipleChoiceOptions;
-        return string.Empty;
-    }
-    private static void AssertQuestions<T>(List<T> questions) where T : IQuestion
-    {
-        if (questions == null)
-        {
-            throw new ArgumentNullException(nameof(questions), "Questions list cannot be null");
-        }
-
-        foreach (var question in questions)
-        {
-            if (question == null)
-            {
-                throw new ArgumentNullException(nameof(question), "Individual question in the list cannot be null");
-            }
-
-            if (question.QuestionNumber < 0)
-            {
-                throw new ArgumentException("Invalid QuestionNumber found in the list");
-            }
-        }
     }
 
     public List<T> DeleteQuestion<T>(List<T> questions, int questionNumber) where T : IQuestion
     {
-        AssertQuestions(questions);
+        ErrorService.AssertQuestions(questions);
 
         int index = GetQuestionNumberIndex(questions, questionNumber);
         if (index != -1)
@@ -347,24 +209,5 @@ public class TestService
 
         return questions;
     }
-
-    public static bool ContatinsInvalidCharacters(string str)
-    {
-        return str.Contains(ErrorStore.IllegalCharacters);
-    }
-
-    public static bool IsValidHz(int hz)
-    {
-        return hz >= 125 && hz <= 8000;
-    }
-
-    public static bool IsValidDecibel(int decibel)
-    {
-        return decibel >= 0 && decibel <= 120;
-    }
-
-    public static bool IsEmptyString(string str)
-    {
-        return string.IsNullOrEmpty(str);
-    }
+    #endregion
 }
